Move testcases into separate file.
